Don't allow both --dir & --git. Use --pick instead

+++ src/fsutil.rs
@@ -1,6 +1,26 @@
-use std::{fs, io, path::Path};
+use std::{fs, path::Path};
 
-pub(crate) fn copy_dir_all(src: impl AsRef<Path>, dst: impl AsRef<Path>) -> io::Result<()> {
+use anyhow::{bail, Result};
+use console::style;
+
+pub(crate) fn copy_all(src: impl AsRef<Path>, dst: impl AsRef<Path>) -> Result<()> {
+    let src_type = fs::metadata(&src)?.file_type();
+    if src_type.is_dir() {
+        copy_dir_all(src, dst)?;
+    } else if src_type.is_file() {
+        fs::create_dir_all(&dst)?;
+        fs::copy(&src, dst.as_ref().join(src.as_ref().file_name().unwrap()))?;
+    } else {
+        bail!(
+            "{} {}",
+            crate::emoji::WARN,
+            style("Symbolic links not supported").bold().red(),
+        )
+    }
+    Ok(())
+}
+
+pub(crate) fn copy_dir_all(src: impl AsRef<Path>, dst: impl AsRef<Path>) -> Result<()> {
     fs::create_dir_all(&dst)?;
     for entry in fs::read_dir(src)? {
         let entry = entry?;
@@ -9,8 +29,14 @@ pub(crate) fn copy_dir_all(src: impl AsRef<Path>, dst: impl AsRef<Path>) -> io::
             if entry.file_name() != ".git" {
                 copy_dir_all(entry.path(), dst.as_ref().join(entry.file_name()))?;
             }
-        } else {
+        } else if ty.is_file() {
             fs::copy(entry.path(), dst.as_ref().join(entry.file_name()))?;
+        } else {
+            bail!(
+                "{} {}",
+                crate::emoji::WARN,
+                style("Symbolic links not supported").bold().red(),
+            )
         }
     }
     Ok(())
